Sunglasses_Prediction/Service1/requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
scikit-learn==1.3.0
pillow==10.0.0   
tensorflow==2.13.0
flask-cors==4.0.0  